Moved all built-in functions
