+++ src/built_ins/mod.rs
@@ -211,7 +211,7 @@ impl BuiltInSub {
 pub mod parser {
     use crate::common::*;
     use crate::parser::pc::*;
-    use crate::parser::Statement;
+    use crate::parser::{Expression, Statement};
 
     /// Parses built-in subs which have a special syntax.
     pub fn parse<R>() -> impl Parser<R, Output = Statement>
@@ -229,10 +229,19 @@ pub mod parser {
             .or(crate::built_ins::put::parser::parse())
             .or(crate::built_ins::view_print::parser::parse())
     }
+
+    // needed for built-in functions that are also keywords (e.g. LEN), so they
+    // cannot be parsed by the `word` module.
+    pub fn built_in_function_call_p<R>() -> impl Parser<R, Output = Expression>
+    where
+        R: Reader<Item = char, Err = QError> + HasLocation + 'static,
+    {
+        crate::built_ins::len::parser::parse()
+    }
 }
 
 pub mod linter {
-    use crate::built_ins::BuiltInSub;
+    use crate::built_ins::{BuiltInFunction, BuiltInSub};
     use crate::common::QErrorNode;
     use crate::parser::ExpressionNodes;
 
@@ -255,10 +264,28 @@ pub mod linter {
             BuiltInSub::ViewPrint => crate::built_ins::view_print::linter::lint(args),
         }
     }
+
+    pub fn lint_function_call(
+        built_in: &BuiltInFunction,
+        args: &ExpressionNodes,
+    ) -> Result<(), QErrorNode> {
+        match built_in {
+            BuiltInFunction::Chr => crate::built_ins::chr::linter::lint(args),
+            BuiltInFunction::Environ => crate::built_ins::environ_fn::linter::lint(args),
+            BuiltInFunction::Eof => crate::built_ins::eof::linter::lint(args),
+            BuiltInFunction::InStr => crate::built_ins::instr::linter::lint(args),
+            BuiltInFunction::LBound => crate::built_ins::lbound::linter::lint(args),
+            BuiltInFunction::Len => crate::built_ins::len::linter::lint(args),
+            BuiltInFunction::Mid => crate::built_ins::mid_fn::linter::lint(args),
+            BuiltInFunction::Str => crate::built_ins::str_fn::linter::lint(args),
+            BuiltInFunction::Val => crate::built_ins::val::linter::lint(args),
+            BuiltInFunction::UBound => crate::built_ins::ubound::linter::lint(args),
+        }
+    }
 }
 
 pub mod interpreter {
-    use crate::built_ins::BuiltInSub;
+    use crate::built_ins::{BuiltInFunction, BuiltInSub};
     use crate::common::QError;
     use crate::interpreter::interpreter_trait::InterpreterTrait;
 
@@ -278,18 +305,45 @@ pub mod interpreter {
             BuiltInSub::ViewPrint => crate::built_ins::view_print::interpreter::run(interpreter),
         }
     }
+
+    pub fn run_function<S: InterpreterTrait>(
+        f: &BuiltInFunction,
+        interpreter: &mut S,
+    ) -> Result<(), QError> {
+        match f {
+            BuiltInFunction::Chr => crate::built_ins::chr::interpreter::run(interpreter),
+            BuiltInFunction::Environ => crate::built_ins::environ_fn::interpreter::run(interpreter),
+            BuiltInFunction::Eof => crate::built_ins::eof::interpreter::run(interpreter),
+            BuiltInFunction::InStr => crate::built_ins::instr::interpreter::run(interpreter),
+            BuiltInFunction::LBound => crate::built_ins::lbound::interpreter::run(interpreter),
+            BuiltInFunction::Len => crate::built_ins::len::interpreter::run(interpreter),
+            BuiltInFunction::Mid => crate::built_ins::mid_fn::interpreter::run(interpreter),
+            BuiltInFunction::Str => crate::built_ins::str_fn::interpreter::run(interpreter),
+            BuiltInFunction::UBound => crate::built_ins::ubound::interpreter::run(interpreter),
+            BuiltInFunction::Val => crate::built_ins::val::interpreter::run(interpreter),
+        }
+    }
 }
 
-pub mod chr;
+mod chr;
 mod close;
+mod environ_fn;
 mod environ_sub;
+mod eof;
 mod field;
 mod get;
 mod input;
+mod instr;
 mod kill;
+mod lbound;
+mod len;
 mod line_input;
 mod lset;
+mod mid_fn;
 mod name;
 mod open;
 mod put;
+mod str_fn;
+mod ubound;
+mod val;
 mod view_print;
