+++ src/interpreter/mod.rs
@@ -1,5 +1,4 @@
 mod arguments;
-mod built_ins;
 pub mod context;
 mod default_stdlib;
 pub mod interpreter;
